fix(ui): Update navBar mobile
